Th2

+++ VirtualCrit IOS/VirtualCrit IOS/Model.swift
@@ -35,7 +35,7 @@ struct Device {
     static var travelDistance                      : Double?
     static var wheelCircumference                  : Double?
     
-    static var maxHR                                : Int = 185
+    static var maxHR                                : Double = 185
     
 
 }
